Corrected normalized detection/tracklets & Tracklets published with minimum confidence threshold value

+++ depthai_bridge/src/ImgDetectionConverter.cpp
@@ -34,14 +34,18 @@ void ImgDetectionConverter::toRosMsg(std::shared_ptr<dai::ImgDetections> inNetDa
 
     // TODO(Sachin): check if this works fine for normalized detection
     // publishing
-    for(int i = 0; i < inNetData->detections.size(); ++i) {
-        int xMin, yMin, xMax, yMax;
-        if(_normalized) {
+    for (int i = 0; i < inNetData->detections.size(); ++i)
+	{
+        float xMin, yMin, xMax, yMax;
+        if(_normalized)
+		{
             xMin = inNetData->detections[i].xmin;
             yMin = inNetData->detections[i].ymin;
             xMax = inNetData->detections[i].xmax;
             yMax = inNetData->detections[i].ymax;
-        } else {
+        }
+		else
+		{
             xMin = inNetData->detections[i].xmin * _width;
             yMin = inNetData->detections[i].ymin * _height;
             xMax = inNetData->detections[i].xmax * _width;
